updated deploy scripts

+++ api-server/deploy.sh
@@ -26,12 +26,21 @@ elif [[ "staging" = $1 ]]; then
   SERVER='staging-api.infinite.industries'  #not used yet
 
   ssh $USER@$SERVER bash --login -i << EOF
-
+    mkdir -p $ROOT/temp-infinite/infinite
     cd $ROOT/temp-infinite/infinite
+
+    if [ -d "./.git" ]
+    then
+      echo "cloning repository"
+      git clone https://github.com/infinite-industries/infinite.git ./
+    else
+      echo 'Updating sources'
+      git reset --hard HEAD
+      git pull origin master
+    fi
+
     echo 'Updating sources'
-    git reset --hard HEAD
-    git checkout development
-    git pull
+    git checkout master
 
     cd ./api-server
     cp * -r $ROOT/infinite/.
@@ -40,11 +49,15 @@ elif [[ "staging" = $1 ]]; then
     echo 'Installing npm packages'
     npm install --production
     forever stop infinite
-    rm $ROOT/.forever/infinite.log
-    forever start --uid infinite index.js
 
-    echo 'Done!'
+    if [ -f "$ROOT/.forever/infinite.log" ]
+    then
+      rm $ROOT/.forever/infinite.log
+    fi
+      forever start --uid infinite index.js
+      echo 'Done!'
 EOF
+
 else
   echo Please specify environment to deploy to.
   echo Usage: ./deploy.sh environment
