+++ web-portal/deploy.sh
@@ -30,32 +30,55 @@ elif [[ "staging" = $1 ]]; then
   ROOT='/home/ubuntu'
   SERVER='staging.infinite.industries'
 
+  npm run production-build
+
   ssh $USER@$SERVER bash --login -i  << EOF
 
+  echo "ROOT: $ROOT"
+  mkdir -p $ROOT/temp-in
   cd $ROOT/temp-infinite
-  echo 'Updating sources'
-  git reset --hard HEAD
-  #git checkout development
-  #git pull
 
-  git pull origin development
+  if [ -d "./.git" ]
+  then
+    echo "cloning repository"
+    git clone https://github.com/infinite-industries/infinite.git ./
+  else
+    echo 'Updating sources'
+    git reset --hard HEAD
+    git pull origin development
+  fi
 
+  git checkout origin development
+
+  mkdir -p $ROOT/front_end_infinite/web-portal/public
   cd ./web-portal
   cp * -r $ROOT/front_end_infinite/.
   cd $ROOT/front_end_infinite
 
+  if [ -d "$ROOT/front_end_infinite/public" ]
+  then
+    rm -R $ROOT/front_end_infinite/public
+  fi
+
   echo 'Installing npm packages'
-  npm install                     # this is super hacky since it brings giant crap like cypress to the server :(
-  #npm install --production
-  echo 'Building frontend js'
-  npm run production-build
+  npm install --production # this is super hacky since it brings giant crap like cypress to the server :(
 
   echo 'Restarting'
   forever stop infinite
-  rm $ROOT/.forever/infinite.log
+
+  if [ -f "$ROOT/.forever/infinite.log" ]
+  then
+    rm $ROOT/.forever/infinite.log
+  fi
 
   forever start --uid infinite ./server.js
   echo 'Done!'
+EOF
+
+  sftp $USER@$SERVER << EOF
+  mkdir $ROOT/front_end_infinite/public
+  cd $ROOT/front_end_infinite/
+  put -r public
 EOF
 else
   echo Please specify environment to deploy to.
